grouper changes
